fix(tool-editor): JSON schema validation is slow and unresponsive
Resolve the code when the user is in mode other than Code mode (this would be done only once when the user switches to that mode) and to not resolve at all while the user is typing, unless the user explicitly clicks a "Resolve"/"Sync" button. Once the user starts typing again, the code will again be unresolved.